Quality-of-life improvements to the WixProject template
EmbedCab is now turned on by default (so that a self-contained MSI file is
generated by the build, instead of a MSI file and a CAB file that must always
be distributed together for the installation to work). I also tweaked the
MajorUpdate logic to rollback the removal of the old version if the upgrade
process fails.